feat: UEC compilance migration and module reorganisation
