+++ packages/devqubit-braket/src/devqubit_braket/adapter.py
@@ -6,7 +6,7 @@ Braket adapter for devqubit tracking system.
 
 Provides integration with Amazon Braket devices, enabling automatic tracking
 of quantum circuit execution, results, and device configurations using the
-Uniform Execution Contract (UEC).
+Uniform Execution Contract (UEC) 1.0.
 
 Example
 -------
@@ -25,45 +25,28 @@ Example
 from __future__ import annotations
 
 import hashlib
-import inspect
 import logging
 from dataclasses import dataclass, field
 from typing import Any
 
-from devqubit_braket.results import extract_counts_payload
-from devqubit_braket.serialization import (
-    BraketCircuitSerializer,
-    circuits_to_text,
-    is_braket_circuit,
-    serialize_openqasm,
+from devqubit_braket.envelope import (
+    create_envelope,
+    log_submission_failure,
 )
-from devqubit_braket.snapshot import create_device_snapshot
-from devqubit_braket.utils import braket_version, extract_task_id, get_backend_name
-from devqubit_engine.circuit.models import CircuitFormat
+from devqubit_braket.serialization import is_braket_circuit
+from devqubit_braket.tracked import TrackedTask, TrackedTaskBatch
+from devqubit_braket.utils import extract_task_id, get_backend_name
 from devqubit_engine.core.run import Run
-from devqubit_engine.uec.device import DeviceSnapshot
 from devqubit_engine.uec.envelope import ExecutionEnvelope
-from devqubit_engine.uec.execution import ExecutionSnapshot
-from devqubit_engine.uec.program import (
-    ProgramArtifact,
-    ProgramSnapshot,
-    TranspilationInfo,
-)
-from devqubit_engine.uec.result import NormalizedCounts, ResultSnapshot
-from devqubit_engine.uec.types import (
-    ArtifactRef,
-    ProgramRole,
-    ResultType,
-    TranspilationMode,
-)
 from devqubit_engine.utils.serialization import to_jsonable
 from devqubit_engine.utils.time_utils import utc_now_iso
 
 
 logger = logging.getLogger(__name__)
 
-# Module-level serializer instance
-_serializer = BraketCircuitSerializer()
+# ============================================================================
+# ProgramSet handling utilities
+# ============================================================================
 
 
 def _is_program_set(obj: Any) -> bool:
@@ -215,6 +198,11 @@ def _materialize_task_spec(
         return task_specification, [task_specification], True, None
 
 
+# ============================================================================
+# Circuit hashing
+# ============================================================================
+
+
 def _compute_circuit_hash(circuits: list[Any]) -> str | None:
     """
     Compute a content hash for circuits.
@@ -287,810 +275,9 @@ def _compute_circuit_hash(circuits: list[Any]) -> str | None:
     return f"sha256:{hashlib.sha256(payload).hexdigest()}"
 
 
-def _serialize_and_log_circuits(
-    tracker: Run,
-    circuits: list[Any],
-    device_name: str,
-) -> list[ArtifactRef]:
-    """
-    Serialize circuits and log as artifacts.
-
-    Logs both JAQCD and OpenQASM formats for comprehensive coverage.
-
-    Parameters
-    ----------
-    tracker : Run
-        Tracker instance.
-    circuits : list
-        List of Braket circuits.
-    device_name : str
-        Backend name for metadata.
-
-    Returns
-    -------
-    list of ArtifactRef
-        References to logged circuit artifacts.
-    """
-    artifact_refs: list[ArtifactRef] = []
-    meta = {
-        "backend_name": device_name,
-        "braket_version": braket_version(),
-    }
-
-    for i, circuit in enumerate(circuits):
-        # Serialize JAQCD (native format)
-        try:
-            jaqcd_data = _serializer.serialize(circuit, CircuitFormat.JAQCD, index=i)
-            ref = tracker.log_bytes(
-                kind="braket.ir.jaqcd",
-                data=jaqcd_data.as_bytes(),
-                media_type="application/json",
-                role="program",
-                meta={**meta, "index": i},
-            )
-            if ref:
-                artifact_refs.append(ref)
-        except Exception as e:
-            logger.debug("Failed to serialize circuit %d to JAQCD: %s", i, e)
-
-        # Serialize OpenQASM (canonical format, better for diffing)
-        try:
-            qasm_data = serialize_openqasm(circuit, index=i)
-            tracker.log_bytes(
-                kind="braket.ir.openqasm",
-                data=qasm_data.as_bytes(),
-                media_type="text/x-qasm; charset=utf-8",
-                role="program",
-                meta={**meta, "index": i, "format": "openqasm3"},
-            )
-        except Exception as e:
-            logger.debug("Failed to serialize circuit %d to OpenQASM: %s", i, e)
-
-    # Log circuit diagrams (human-readable)
-    try:
-        diagram_text = circuits_to_text(circuits)
-        tracker.log_bytes(
-            kind="braket.circuits.diagram",
-            data=diagram_text.encode("utf-8"),
-            media_type="text/plain; charset=utf-8",
-            role="program",
-            meta={"num_circuits": len(circuits)},
-        )
-    except Exception as e:
-        logger.debug("Failed to generate circuit diagrams: %s", e)
-
-    return artifact_refs
-
-
-def _create_program_snapshot(
-    circuits: list[Any],
-    artifact_refs: list[ArtifactRef],
-    circuit_hash: str | None,
-) -> ProgramSnapshot:
-    """
-    Create a ProgramSnapshot from circuits and their artifact refs.
-
-    Parameters
-    ----------
-    circuits : list
-        List of Braket circuits.
-    artifact_refs : list of ArtifactRef
-        References to logged circuit artifacts.
-    circuit_hash : str or None
-        Circuit structure hash.
-
-    Returns
-    -------
-    ProgramSnapshot
-        Program snapshot with logical artifacts.
-    """
-    logical_artifacts: list[ProgramArtifact] = []
-
-    for i, ref in enumerate(artifact_refs):
-        circuit_name = None
-        if i < len(circuits):
-            circuit_name = getattr(circuits[i], "name", None)
-
-        logical_artifacts.append(
-            ProgramArtifact(
-                ref=ref,
-                role=ProgramRole.LOGICAL,
-                format="jaqcd",
-                name=circuit_name or f"circuit_{i}",
-                index=i,
-            )
-        )
-
-    return ProgramSnapshot(
-        logical=logical_artifacts,
-        physical=[],  # Braket doesn't expose transpiled circuits
-        program_hash=circuit_hash,
-        num_circuits=len(circuits),
-    )
-
-
-def _create_execution_snapshot(
-    shots: int | None,
-    task_ids: list[str],
-    submitted_at: str,
-    options: dict[str, Any] | None = None,
-) -> ExecutionSnapshot:
-    """
-    Create an ExecutionSnapshot for a Braket task submission.
-
-    Parameters
-    ----------
-    shots : int or None
-        Number of shots (None means provider default).
-    task_ids : list of str
-        Task identifiers.
-    submitted_at : str
-        ISO 8601 submission timestamp.
-    options : dict, optional
-        Additional execution options.
-
-    Returns
-    -------
-    ExecutionSnapshot
-        Execution metadata snapshot.
-    """
-    return ExecutionSnapshot(
-        submitted_at=submitted_at,
-        shots=shots,
-        task_ids=task_ids,
-        execution_count=len(task_ids) if task_ids else 1,
-        transpilation=TranspilationInfo(
-            mode=TranspilationMode.MANAGED,
-            transpiled_by="provider",
-        ),
-        options=options or {},
-        sdk="braket",
-    )
-
-
-def _create_result_snapshot(
-    result: Any,
-    raw_result_ref: ArtifactRef | None,
-    shots: int | None,
-    error_message: str | None = None,
-) -> ResultSnapshot:
-    """
-    Create a ResultSnapshot from Braket result.
-
-    Parameters
-    ----------
-    result : Any
-        Braket result object (may be None on failure).
-    raw_result_ref : ArtifactRef or None
-        Reference to raw result artifact.
-    shots : int or None
-        Number of shots used.
-    error_message : str or None
-        Error message if execution failed.
-
-    Returns
-    -------
-    ResultSnapshot
-        Result snapshot with normalized counts and success status.
-    """
-    normalized_counts: list[NormalizedCounts] = []
-    success = False
-    result_type = ResultType.COUNTS
-
-    if result is not None and error_message is None:
-        # Check if result is already a combined payload dict (from batch)
-        if isinstance(result, dict) and "experiments" in result:
-            counts_payload = result
-        else:
-            counts_payload = extract_counts_payload(result)
-
-        if counts_payload and counts_payload.get("experiments"):
-            for exp in counts_payload["experiments"]:
-                counts = exp.get("counts", {})
-                normalized_counts.append(
-                    NormalizedCounts(
-                        circuit_index=exp.get("index", 0),
-                        counts=counts if counts else {},
-                        shots=shots,
-                        name=exp.get("name"),
-                    )
-                )
-
-        # Fallback: if we have a result but no experiments extracted
-        if not normalized_counts:
-            batch_size = result.get("batch_size", 1) if isinstance(result, dict) else 1
-            for i in range(batch_size):
-                normalized_counts.append(
-                    NormalizedCounts(circuit_index=i, counts={}, shots=shots)
-                )
-
-        # Success = we have actual non-empty counts
-        success = any(nc.counts for nc in normalized_counts)
-
-        # For shots=0 (analytical), may get statevector/other instead of counts
-        if not success and shots == 0:
-            if hasattr(result, "values") or hasattr(result, "result_types"):
-                result_type = ResultType.OTHER
-                success = True
-
-    # Build ResultSnapshot - handle error_message field defensively
-    snapshot_kwargs: dict[str, Any] = {
-        "result_type": result_type,
-        "raw_result_ref": raw_result_ref,
-        "counts": normalized_counts,
-        "num_experiments": len(normalized_counts),
-        "success": success,
-    }
-
-    # Only add error_message if the field exists in ResultSnapshot
-    try:
-        sig = inspect.signature(ResultSnapshot)
-        if "error_message" in sig.parameters:
-            snapshot_kwargs["error_message"] = error_message
-    except Exception:
-        pass
-
-    return ResultSnapshot(**snapshot_kwargs)
-
-
-def _create_envelope(
-    tracker: Run,
-    device: Any,
-    circuits: list[Any],
-    shots: int | None,
-    task_ids: list[str],
-    submitted_at: str,
-    circuit_hash: str | None,
-    options: dict[str, Any] | None = None,
-) -> ExecutionEnvelope:
-    """
-    Create and log a complete ExecutionEnvelope (pre-result).
-
-    Parameters
-    ----------
-    tracker : Run
-        Tracker instance.
-    device : Any
-        Braket device.
-    circuits : list
-        List of circuits.
-    shots : int or None
-        Number of shots.
-    task_ids : list of str
-        Task identifiers.
-    submitted_at : str
-        Submission timestamp.
-    circuit_hash : str or None
-        Circuit hash.
-    options : dict, optional
-        Execution options.
-
-    Returns
-    -------
-    ExecutionEnvelope
-        Envelope with device, program, and execution snapshots.
-    """
-    device_name = get_backend_name(device=device)
-
-    # Create device snapshot with tracker for raw_properties logging
-    try:
-        device_snapshot = create_device_snapshot(device=device, tracker=tracker)
-    except Exception as e:
-        logger.warning(
-            "Failed to create device snapshot: %s. Using minimal snapshot.", e
-        )
-        # Create minimal snapshot on failure
-        device_snapshot = DeviceSnapshot(
-            captured_at=utc_now_iso(),
-            backend_name=device_name,
-            backend_type="unknown",
-            provider="braket",
-            sdk_versions={"braket": braket_version()},
-        )
-
-    # Update tracker record
-    tracker.record["device_snapshot"] = {
-        "sdk": "braket",
-        "backend_name": device_name,
-        "backend_type": device_snapshot.backend_type,
-        "provider": device_snapshot.provider,
-        "captured_at": device_snapshot.captured_at,
-        "num_qubits": device_snapshot.num_qubits,
-        "calibration_summary": device_snapshot.get_calibration_summary(),
-    }
-
-    # Log circuits and get artifact refs
-    artifact_refs = _serialize_and_log_circuits(
-        tracker=tracker,
-        circuits=circuits,
-        device_name=device_name,
-    )
-
-    # Create program snapshot
-    program_snapshot = _create_program_snapshot(
-        circuits=circuits,
-        artifact_refs=artifact_refs,
-        circuit_hash=circuit_hash,
-    )
-
-    # Create execution snapshot
-    execution_snapshot = _create_execution_snapshot(
-        shots=shots,
-        task_ids=task_ids,
-        submitted_at=submitted_at,
-        options=options,
-    )
-
-    return ExecutionEnvelope(
-        schema_version="devqubit.envelope/0.1",
-        adapter="braket",
-        created_at=utc_now_iso(),
-        device=device_snapshot,
-        program=program_snapshot,
-        execution=execution_snapshot,
-        result=None,  # Will be filled when result() is called
-    )
-
-
-def _finalize_envelope_with_result(
-    tracker: Run,
-    envelope: ExecutionEnvelope,
-    result: Any,
-    device_name: str,
-    shots: int | None,
-    error_message: str | None = None,
-) -> ExecutionEnvelope:
-    """
-    Finalize envelope with result and log it.
-
-    This function never raises exceptions - tracking should never crash
-    user experiments. Validation errors are logged but execution continues.
-
-    Parameters
-    ----------
-    tracker : Run
-        Tracker instance.
-    envelope : ExecutionEnvelope
-        Envelope to finalize.
-    result : Any
-        Braket result object (may be None on failure).
-    device_name : str
-        Device name.
-    shots : int or None
-        Number of shots.
-    error_message : str or None
-        Error message if execution failed.
-
-    Returns
-    -------
-    ExecutionEnvelope
-        Finalized envelope.
-
-    Raises
-    ------
-    ValueError
-        If envelope is None.
-    """
-    if envelope is None:
-        raise ValueError("Cannot finalize None envelope")
-
-    # Log raw result and get ref
-    raw_result_ref: ArtifactRef | None = None
-    if result is not None:
-        try:
-            result_payload = to_jsonable(result)
-        except Exception:
-            result_payload = {"repr": repr(result)[:2000]}
-
-        try:
-            raw_result_ref = tracker.log_json(
-                name="braket.result",
-                obj=result_payload,
-                role="results",
-                kind="result.braket.raw.json",
-            )
-        except Exception as e:
-            logger.warning("Failed to log raw result: %s", e)
-    elif error_message:
-        try:
-            tracker.log_json(
-                name="braket.error",
-                obj={"error": error_message, "timestamp": utc_now_iso()},
-                role="results",
-                kind="result.braket.error.json",
-            )
-        except Exception as e:
-            logger.warning("Failed to log error: %s", e)
-
-    # Create result snapshot
-    result_snapshot = _create_result_snapshot(
-        result, raw_result_ref, shots, error_message
-    )
-
-    # Update execution snapshot with completion time
-    if envelope.execution:
-        envelope.execution.completed_at = utc_now_iso()
-
-    # Add result to envelope
-    envelope.result = result_snapshot
-
-    # Extract counts for separate logging
-    counts_payload = None
-    if result is not None:
-        try:
-            counts_payload = extract_counts_payload(result)
-        except Exception as e:
-            logger.debug("Failed to extract counts payload: %s", e)
-
-    # Validate and log envelope
-    try:
-        tracker.log_envelope(envelope=envelope)
-    except Exception as e:
-        logger.warning("Failed to log envelope: %s", e)
-
-    # Log normalized counts
-    if counts_payload is not None:
-        try:
-            tracker.log_json(
-                name="counts",
-                obj=counts_payload,
-                role="results",
-                kind="result.counts.json",
-            )
-        except Exception as e:
-            logger.debug("Failed to log counts: %s", e)
-
-    # Update tracker record
-    tracker.record["results"] = {
-        "completed_at": utc_now_iso(),
-        "backend_name": device_name,
-        "num_experiments": result_snapshot.num_experiments,
-        "result_type": result_snapshot.result_type.value,
-        "success": result_snapshot.success,
-    }
-    if error_message:
-        tracker.record["results"]["error"] = error_message
-
-    logger.debug("Logged execution envelope for %s", device_name)
-
-    return envelope
-
-
-def _log_submission_failure(
-    tracker: Run,
-    device_name: str,
-    error: Exception,
-    circuits: list[Any],
-    shots: int | None,
-    submitted_at: str,
-) -> None:
-    """
-    Log a task submission failure.
-
-    Parameters
-    ----------
-    tracker : Run
-        Tracker instance.
-    device_name : str
-        Device name.
-    error : Exception
-        The exception that occurred.
-    circuits : list
-        Circuits that were being submitted.
-    shots : int or None
-        Requested shots.
-    submitted_at : str
-        Submission timestamp.
-    """
-    error_info = {
-        "type": "submission_failure",
-        "error_type": type(error).__name__,
-        "error_message": str(error),
-        "device_name": device_name,
-        "num_circuits": len(circuits),
-        "shots": shots,
-        "submitted_at": submitted_at,
-        "failed_at": utc_now_iso(),
-    }
-
-    tracker.log_json(
-        name="submission_failure",
-        obj=error_info,
-        role="error",
-        kind="devqubit.submission_failure.json",
-    )
-
-    tracker.record["submission_failure"] = error_info
-    logger.warning("Task submission failed on %s: %s", device_name, error)
-
-
-def _combine_batch_results(results_list: list[Any]) -> dict[str, Any]:
-    """
-    Combine batch results into a single structure for logging.
-
-    Parameters
-    ----------
-    results_list : list
-        List of individual result objects.
-
-    Returns
-    -------
-    dict
-        Combined result structure.
-    """
-    experiments: list[dict[str, Any]] = []
-
-    for i, result in enumerate(results_list):
-        if result is None:
-            experiments.append({"index": i, "status": "failed", "counts": {}})
-            continue
-
-        counts_payload = extract_counts_payload(result)
-        if counts_payload and counts_payload.get("experiments"):
-            for exp in counts_payload["experiments"]:
-                exp_copy = dict(exp)
-                exp_copy["batch_index"] = i
-                experiments.append(exp_copy)
-        else:
-            experiments.append({"index": i, "batch_index": i, "counts": {}})
-
-    return {"experiments": experiments, "batch_size": len(results_list)}
-
-
-@dataclass
-class TrackedTask:
-    """
-    Wrapper for Braket task that tracks result retrieval.
-
-    Intercepts `result()` calls to finalize the execution envelope
-    with result data. Handles exceptions gracefully with failure logging.
-
-    Parameters
-    ----------
-    task : Any
-        Original Braket task instance.
-    tracker : Run
-        Tracker instance for logging.
-    device_name : str
-        Name of the device that created this task.
-    envelope : ExecutionEnvelope or None
-        Envelope to finalize with results.
-    shots : int or None
-        Number of shots for this execution.
-    should_log_results : bool
-        Whether to log results for this task.
-    """
-
-    task: Any
-    tracker: Run
-    device_name: str
-    envelope: ExecutionEnvelope | None = None
-    shots: int | None = None
-    should_log_results: bool = True
-    _result_logged: bool = field(default=False, init=False, repr=False)
-
-    def result(self, *args: Any, **kwargs: Any) -> Any:
-        """
-        Retrieve task result and finalize envelope.
-
-        Handles exceptions gracefully, logging failure information
-        before re-raising.
-
-        Parameters
-        ----------
-        *args : Any
-            Positional arguments passed to underlying result().
-        **kwargs : Any
-            Keyword arguments passed to underlying result().
-
-        Returns
-        -------
-        Any
-            Braket result object.
-
-        Raises
-        ------
-        Exception
-            Re-raises any exception from underlying result() after logging.
-        """
-        result = None
-        error_message: str | None = None
-
-        try:
-            result = self.task.result(*args, **kwargs)
-        except Exception as e:
-            error_message = f"{type(e).__name__}: {e}"
-            logger.warning("Task result() failed on %s: %s", self.device_name, e)
-
-            # Log failure even if we re-raise
-            if self.should_log_results and self.envelope and not self._result_logged:
-                self._result_logged = True
-                try:
-                    _finalize_envelope_with_result(
-                        tracker=self.tracker,
-                        envelope=self.envelope,
-                        result=None,
-                        device_name=self.device_name,
-                        shots=self.shots,
-                        error_message=error_message,
-                    )
-                except Exception as log_err:
-                    logger.warning(
-                        "Failed to log error envelope for %s: %s",
-                        self.device_name,
-                        log_err,
-                    )
-            raise
-
-        # Log successful result
-        if self.should_log_results and self.envelope and not self._result_logged:
-            self._result_logged = True
-            try:
-                _finalize_envelope_with_result(
-                    tracker=self.tracker,
-                    envelope=self.envelope,
-                    result=result,
-                    device_name=self.device_name,
-                    shots=self.shots,
-                )
-                logger.debug("Finalized envelope for task on %s", self.device_name)
-            except Exception as log_err:
-                logger.warning(
-                    "Failed to finalize envelope for %s: %s",
-                    self.device_name,
-                    log_err,
-                )
-                # Record error in tracker for visibility
-                self.tracker.record.setdefault("warnings", []).append(
-                    {
-                        "type": "result_logging_failed",
-                        "message": str(log_err),
-                        "device_name": self.device_name,
-                    }
-                )
-
-        return result
-
-    def __getattr__(self, name: str) -> Any:
-        """Delegate attribute access to wrapped task."""
-        return getattr(self.task, name)
-
-    def __repr__(self) -> str:
-        """Return string representation."""
-        task_id = extract_task_id(self.task) or "unknown"
-        return f"TrackedTask(device={self.device_name!r}, task_id={task_id!r})"
-
-
-@dataclass
-class TrackedTaskBatch:
-    """
-    Wrapper for Braket task batch that tracks result retrieval.
-
-    Wraps AwsQuantumTaskBatch to intercept `results()` calls and log
-    all results with proper handling of partial failures.
-
-    Parameters
-    ----------
-    batch : Any
-        Original Braket task batch instance.
-    tracker : Run
-        Tracker instance for logging.
-    device_name : str
-        Name of the device that created this batch.
-    envelope : ExecutionEnvelope or None
-        Envelope to finalize with results.
-    shots : int or None
-        Number of shots for this execution.
-    should_log_results : bool
-        Whether to log results for this batch.
-    """
-
-    batch: Any
-    tracker: Run
-    device_name: str
-    envelope: ExecutionEnvelope | None = None
-    shots: int | None = None
-    should_log_results: bool = True
-    _results_logged: bool = field(default=False, init=False, repr=False)
-
-    def results(self, *args: Any, **kwargs: Any) -> list[Any]:
-        """
-        Retrieve batch results and finalize envelope.
-
-        Handles partial failures (None results for failed tasks).
-
-        Parameters
-        ----------
-        *args : Any
-            Positional arguments passed to underlying results().
-        **kwargs : Any
-            Keyword arguments passed to underlying results().
-
-        Returns
-        -------
-        list
-            List of Braket result objects (may contain None for failed tasks).
-        """
-        results_list: list[Any] = []
-        error_message: str | None = None
-
-        try:
-            results_list = self.batch.results(*args, **kwargs)
-        except Exception as e:
-            error_message = f"{type(e).__name__}: {e}"
-            logger.warning("Batch results() failed on %s: %s", self.device_name, e)
-
-            if self.should_log_results and self.envelope and not self._results_logged:
-                self._results_logged = True
-                try:
-                    _finalize_envelope_with_result(
-                        self.tracker,
-                        self.envelope,
-                        None,
-                        self.device_name,
-                        self.shots,
-                        error_message=error_message,
-                    )
-                except Exception as log_err:
-                    logger.warning(
-                        "Failed to log error envelope for batch %s: %s",
-                        self.device_name,
-                        log_err,
-                    )
-            raise
-
-        # Check for partial failures (None in results)
-        failed_count = sum(1 for r in results_list if r is None)
-        if failed_count > 0:
-            logger.warning(
-                "Batch on %s: %d/%d tasks failed",
-                self.device_name,
-                failed_count,
-                len(results_list),
-            )
-            error_message = (
-                f"Partial failure: {failed_count}/{len(results_list)} tasks failed"
-            )
-
-        # Aggregate successful results for logging
-        if self.should_log_results and self.envelope and not self._results_logged:
-            self._results_logged = True
-            try:
-                combined_result = _combine_batch_results(results_list)
-
-                _finalize_envelope_with_result(
-                    self.tracker,
-                    self.envelope,
-                    combined_result,
-                    self.device_name,
-                    self.shots,
-                    error_message=error_message if failed_count > 0 else None,
-                )
-                logger.debug("Finalized envelope for batch on %s", self.device_name)
-            except Exception as log_err:
-                logger.warning(
-                    "Failed to finalize envelope for batch %s: %s",
-                    self.device_name,
-                    log_err,
-                )
-                # Record error in tracker for visibility
-                self.tracker.record.setdefault("warnings", []).append(
-                    {
-                        "type": "batch_result_logging_failed",
-                        "message": str(log_err),
-                        "device_name": self.device_name,
-                    }
-                )
-
-        return results_list
-
-    def __getattr__(self, name: str) -> Any:
-        """Delegate attribute access to wrapped batch."""
-        return getattr(self.batch, name)
-
-    def __repr__(self) -> str:
-        """Return string representation."""
-        return f"TrackedTaskBatch(device={self.device_name!r})"
+# ============================================================================
+# TrackedDevice - wraps Braket devices with tracking
+# ============================================================================
 
 
 @dataclass
@@ -1199,7 +386,7 @@ class TrackedDevice:
                 task = self.device.run(run_payload, shots=shots, *args, **kwargs)
         except Exception as e:
             if should_log and circuits_for_logging:
-                _log_submission_failure(
+                log_submission_failure(
                     self.tracker,
                     device_name,
                     e,
@@ -1216,7 +403,7 @@ class TrackedDevice:
         # Create envelope if logging
         envelope: ExecutionEnvelope | None = None
         if should_log and circuits_for_logging:
-            envelope = _create_envelope(
+            envelope = create_envelope(
                 tracker=self.tracker,
                 device=self.device,
                 circuits=circuits_for_logging,
@@ -1224,6 +411,7 @@ class TrackedDevice:
                 task_ids=task_ids,
                 submitted_at=submitted_at,
                 circuit_hash=circuit_hash,
+                execution_index=exec_count,  # P0 fix: pass actual execution index
                 options=options if options else None,
             )
 
@@ -1232,10 +420,10 @@ class TrackedDevice:
 
             self._logged_execution_count += 1
 
-            # Set tracker tags/params
+            # Set tracker tags/params (P1 fix: provider is platform, not SDK)
             self.tracker.set_tag("backend_name", device_name)
-            self.tracker.set_tag("provider", "braket")
-            self.tracker.set_tag("adapter", "braket")
+            self.tracker.set_tag("provider", "aws_braket")
+            self.tracker.set_tag("adapter", "devqubit-braket")
 
             if shots is not None:
                 self.tracker.log_param("shots", int(shots))
@@ -1245,7 +433,7 @@ class TrackedDevice:
             self.tracker.record["backend"] = {
                 "name": device_name,
                 "type": self.device.__class__.__name__,
-                "provider": "braket",
+                "provider": "aws_braket",  # P1 fix: platform, not SDK
             }
 
             self.tracker.record["execute"] = {
@@ -1346,7 +534,7 @@ class TrackedDevice:
                 )
         except Exception as e:
             if should_log and circuits_for_logging:
-                _log_submission_failure(
+                log_submission_failure(
                     self.tracker,
                     device_name,
                     e,
@@ -1359,7 +547,7 @@ class TrackedDevice:
         # Create envelope if logging
         envelope: ExecutionEnvelope | None = None
         if should_log and circuits_for_logging:
-            envelope = _create_envelope(
+            envelope = create_envelope(
                 tracker=self.tracker,
                 device=self.device,
                 circuits=circuits_for_logging,
@@ -1367,6 +555,7 @@ class TrackedDevice:
                 task_ids=[],  # Batch doesn't have a single ID upfront
                 submitted_at=submitted_at,
                 circuit_hash=circuit_hash,
+                execution_index=exec_count,  # P0 fix: pass actual execution index
                 options=options,
             )
 
@@ -1375,10 +564,10 @@ class TrackedDevice:
 
             self._logged_execution_count += 1
 
-            # Set tracker tags/params
+            # Set tracker tags/params (P1 fix: provider is platform, not SDK)
             self.tracker.set_tag("backend_name", device_name)
-            self.tracker.set_tag("provider", "braket")
-            self.tracker.set_tag("adapter", "braket")
+            self.tracker.set_tag("provider", "aws_braket")
+            self.tracker.set_tag("adapter", "devqubit-braket")
             self.tracker.set_tag("batch_execution", "true")
 
             if shots is not None:
@@ -1390,7 +579,7 @@ class TrackedDevice:
             self.tracker.record["backend"] = {
                 "name": device_name,
                 "type": self.device.__class__.__name__,
-                "provider": "braket",
+                "provider": "aws_braket",  # P1 fix: platform, not SDK
             }
 
             self.tracker.record["execute"] = {
@@ -1459,6 +648,11 @@ class TrackedDevice:
         return f"TrackedDevice(device={device_name!r}, run_id={self.tracker.run_id!r})"
 
 
+# ============================================================================
+# BraketAdapter - main adapter class
+# ============================================================================
+
+
 class BraketAdapter:
     """
     Adapter for integrating Braket devices with devqubit tracking.
@@ -1515,7 +709,7 @@ class BraketAdapter:
         return {
             "name": get_backend_name(device),
             "type": device.__class__.__name__,
-            "provider": "braket",
+            "provider": "aws_braket",  # P1 fix: platform, not SDK
         }
 
     def wrap_executor(
